(refactor) Package structure change
