+++ src/main/java/lotto/domain/WinningPercent.java
@@ -1,8 +1,8 @@
-package lotto;
+package lotto.domain;
 
 import java.util.Map;
 import java.util.Objects;
-import lotto.WinningNumbers.Rank;
+import lotto.domain.WinningNumbers.Rank;
 
 public class WinningPercent {
